Implement for more std types
