+++ crates/borrowme/src/borrow.rs
@@ -1,7 +1,5 @@
 #[cfg(feature = "std")]
-use core::hash::Hash;
-#[cfg(feature = "std")]
-use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet, LinkedList};
+mod std;
 
 /// Borrow from self.
 ///
@@ -90,16 +88,6 @@ pub trait Borrow {
     fn borrow(&self) -> Self::Target<'_>;
 }
 
-#[cfg(feature = "std")]
-impl Borrow for String {
-    type Target<'a> = &'a str;
-
-    #[inline]
-    fn borrow(&self) -> Self::Target<'_> {
-        self.as_str()
-    }
-}
-
 impl<T> Borrow for Option<T>
 where
     T: Borrow,
@@ -120,105 +108,3 @@ impl<T> Borrow for [T] {
         self
     }
 }
-
-macro_rules! seq {
-    (cap $seq:ident, $insert:ident $(, $trait:path)* $(,)?) => {
-        #[cfg(feature = "std")]
-        impl<T> Borrow for $seq<T>
-        where
-            T: Borrow,
-            $(for<'a> T::Target<'a>: $trait,)*
-        {
-            type Target<'a> = $seq<T::Target<'a>> where T: 'a;
-
-            #[inline]
-            fn borrow(&self) -> Self::Target<'_> {
-                let mut out = <$seq<_>>::with_capacity(self.len());
-
-                for value in self {
-                    out.$insert(value.borrow());
-                }
-
-                out
-            }
-        }
-    };
-
-    ($seq:ident, $insert:ident $(, $trait:path)* $(,)?) => {
-        #[cfg(feature = "std")]
-        impl<T> Borrow for $seq<T>
-        where
-            T: Borrow,
-            $(for<'a> T::Target<'a>: $trait,)*
-        {
-            type Target<'a> = $seq<T::Target<'a>> where T: 'a;
-
-            #[inline]
-            fn borrow(&self) -> Self::Target<'_> {
-                let mut out = <$seq<_>>::new();
-
-                for value in self {
-                    out.$insert(value.borrow());
-                }
-
-                out
-            }
-        }
-    };
-}
-
-macro_rules! map {
-    (cap $map:ident, $insert:ident $(, $trait:path)* $(,)?) => {
-        #[cfg(feature = "std")]
-        impl<K, V> Borrow for $map<K, V>
-        where
-            K: Borrow,
-            V: Borrow,
-            $(for<'a> K::Target<'a>: $trait,)*
-        {
-            type Target<'a> = $map<K::Target<'a>, V::Target<'a>> where K: 'a, V: 'a;
-
-            #[inline]
-            fn borrow(&self) -> Self::Target<'_> {
-                let mut out = <$map<_, _>>::with_capacity(self.len());
-
-                for (key, value) in self {
-                    out.$insert(key.borrow(), value.borrow());
-                }
-
-                out
-            }
-        }
-    };
-
-    ($map:ident, $insert:ident $(, $trait:path)* $(,)?) => {
-        #[cfg(feature = "std")]
-        impl<K, V> Borrow for $map<K, V>
-        where
-            K: Borrow,
-            V: Borrow,
-            $(for<'a> K::Target<'a>: $trait,)*
-        {
-            type Target<'a> = $map<K::Target<'a>, V::Target<'a>> where K: 'a, V: 'a;
-
-            #[inline]
-            fn borrow(&self) -> Self::Target<'_> {
-                let mut out = <$map<_, _>>::new();
-
-                for (key, value) in self {
-                    out.$insert(key.borrow(), value.borrow());
-                }
-
-                out
-            }
-        }
-    };
-}
-
-seq!(cap Vec, push);
-seq!(cap HashSet, insert, Hash, Eq);
-seq!(BTreeSet, insert, PartialOrd, Ord, Eq);
-seq!(LinkedList, push_back);
-
-map!(cap HashMap, insert, Hash, Eq);
-map!(BTreeMap, insert, PartialOrd, Ord, Eq);
