+++ crates/borrowme/src/to_owned.rs
@@ -1,7 +1,5 @@
 #[cfg(feature = "std")]
-use core::hash::Hash;
-#[cfg(feature = "std")]
-use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet, LinkedList};
+mod std;
 
 /// Convert to owned.
 ///
@@ -81,140 +79,3 @@ where
         T::to_owned(*self)
     }
 }
-
-#[cfg(feature = "std")]
-impl ToOwned for str {
-    type Owned = String;
-
-    #[inline]
-    fn to_owned(&self) -> Self::Owned {
-        String::from(self)
-    }
-}
-
-impl<T> ToOwned for Option<T>
-where
-    T: ToOwned,
-{
-    type Owned = Option<T::Owned>;
-
-    #[inline]
-    fn to_owned(&self) -> Self::Owned {
-        self.as_ref().map(ToOwned::to_owned)
-    }
-}
-
-#[cfg(feature = "std")]
-impl<T> ToOwned for [T]
-where
-    T: Clone,
-{
-    type Owned = Vec<T>;
-
-    #[inline]
-    fn to_owned(&self) -> Self::Owned {
-        self.to_vec()
-    }
-}
-
-macro_rules! seq {
-    (cap $seq:ident, $insert:ident $(, $trait:path)* $(,)?) => {
-        #[cfg(feature = "std")]
-        impl<T> ToOwned for $seq<T>
-        where
-            T: ToOwned,
-            $(T::Owned: $trait,)*
-        {
-            type Owned = $seq<T::Owned>;
-
-            #[inline]
-            fn to_owned(&self) -> Self::Owned {
-                let mut out = <$seq<T::Owned>>::with_capacity(self.len());
-
-                for value in self.iter() {
-                    out.$insert(value.to_owned());
-                }
-
-                out
-            }
-        }
-    };
-
-    ($seq:ident, $insert:ident $(, $trait:path)* $(,)?) => {
-        #[cfg(feature = "std")]
-        impl<T> ToOwned for $seq<T>
-        where
-            T: ToOwned,
-            $(T::Owned: $trait,)*
-        {
-            type Owned = $seq<T::Owned>;
-
-            #[inline]
-            fn to_owned(&self) -> Self::Owned {
-                let mut out = <$seq<T::Owned>>::new();
-
-                for value in self.iter() {
-                    out.$insert(value.to_owned());
-                }
-
-                out
-            }
-        }
-    };
-}
-
-macro_rules! map {
-    (cap $map:ident, $insert:ident $(, $trait:path)* $(,)?) => {
-        #[cfg(feature = "std")]
-        impl<K, V> ToOwned for $map<K, V>
-        where
-            K: ToOwned,
-            V: ToOwned,
-            $(K::Owned: $trait,)*
-        {
-            type Owned = $map<K::Owned, V::Owned>;
-
-            #[inline]
-            fn to_owned(&self) -> Self::Owned {
-                let mut out = <$map<_, _>>::with_capacity(self.len());
-
-                for (key, value) in self.iter() {
-                    out.$insert(key.to_owned(), value.to_owned());
-                }
-
-                out
-            }
-        }
-    };
-
-    ($map:ident, $insert:ident $(, $trait:path)* $(,)?) => {
-        #[cfg(feature = "std")]
-        impl<K, V> ToOwned for $map<K, V>
-        where
-            K: ToOwned,
-            V: ToOwned,
-            $(K::Owned: $trait,)*
-        {
-            type Owned = $map<K::Owned, V::Owned>;
-
-            #[inline]
-            fn to_owned(&self) -> Self::Owned {
-                let mut out = <$map<_, _>>::new();
-
-                for (key, value) in self.iter() {
-                    out.$insert(key.to_owned(), value.to_owned());
-                }
-
-                out
-            }
-        }
-    };
-}
-
-seq!(cap HashSet, insert, Hash, Eq);
-seq!(cap Vec, push);
-seq!(BTreeSet, insert, PartialOrd, Ord, Eq);
-seq!(LinkedList, push_back);
-
-map!(cap HashMap, insert, Hash, Eq);
-map!(BTreeMap, insert, PartialOrd, Ord, Eq);
